sample_rest_flask/requirements.txt
aniso8601==9.0.1
astroid==2.9.3
click==8.0.3
Flask==2.0.3
Flask-HTTPAuth==4.5.0
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mccabe==0.6.1
passlib==1.7.4
platformdirs==2.5.0
PyJWT==2.3.0
pylint==2.12.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.31
toml==0.10.2
typing-extensions==4.1.1
Werkzeug==2.0.3
wrapt==1.13.3
